+++ src/nustattools/stats/_derate.py
@@ -49,6 +49,9 @@ def fill_max_correlation(cor: ArrayLike, target: ArrayLike) -> NDArray[Any]:
     cora = np.array(cor)
     target = np.asarray(target)
 
+    # Check and fix connections to other elements
+    cora = _fix(cora)
+
     priority = np.unravel_index(
         np.argsort(np.abs(target), axis=None)[::-1], target.shape
     )
@@ -69,16 +72,27 @@ def fill_max_correlation(cor: ArrayLike, target: ArrayLike) -> NDArray[Any]:
 
 
 def get_blocks(cov: NDArray[Any]) -> list[int]:
-    """Determine the sizes of known block matrices."""
+    """Determine the sizes of known block matrices.
+
+    Assumes the matrix is symmetric.
+
+    """
 
     blocks = []
-    n = 0
-    i = 0
-    for j in range(cov.shape[0]):
-        if np.isnan(cov[i, j]):
+    n = 1
+
+    # Find blocks by looking at NaNs
+    nans = np.isnan(cov)
+    trans = np.any(nans[:-1] ^ nans[1:], axis=1)
+
+    # Find blocks by looking at zeros on diagonal
+    zeros = np.diag(cov) == 0
+    trans |= zeros[:-1] ^ zeros[1:]
+
+    for j in range(cov.shape[0] - 1):
+        if trans[j]:
             blocks.append(n)
             n = 1
-            i = j
         else:
             n += 1
 
@@ -96,7 +110,10 @@ def get_whitening_transform(cov: NDArray[Any]) -> NDArray[Any]:
     i = 0
     for n in blocks:
         c = cov[i : i + n, :][:, i : i + n]
-        W_l.append(np.linalg.inv(sqrtm(c)))
+        if np.all(c == 0):
+            W_l.append(np.zeros_like(c))
+        else:
+            W_l.append(np.linalg.inv(sqrtm(c)))
         i += n
 
     return np.asarray(block_diag(*W_l))
@@ -117,7 +134,9 @@ def derate_covariance(
     ----------
     cov : numpy.ndarray or list of numpy.ndarray
         The covariance matrix of the data or a list of covariances that add up
-        to the total. Unknown covariances must be ``np.nan``.
+        to the total. Unknown covariance blocks must be ``np.nan``. Off
+        diagonal blocks may only be ``0'' or ``np.nan''. Diagonal blocks must
+        not be ``np.nan''.
     jacobian : numpy.ndarray, default=None
         Jacobian matrix of the model prediction wrt the best-fit parameters.
     sigma : float, default=3.